rename executor test
